zq update through "zqd listen: add -portfile option" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1017, authored by @mattnibs,
has been merged.

zqd listen: add -portfile option

Add -portfile option that atomically writes the port of the zqd
listener to a file once the listener has been established.

The main purpose for this option is for testing, where port 0 will
be passed instructing the os to select a random available port.

Part of brimsec/zq#975