zq update through "Add changes to zql.es.js" by jameskerr
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1020, authored by @jameskerr,
has been merged.

Add changes to zql.es.js

Soon the app will only rely on the `zql.es.js` version of the parser. It appears some changes went in to the zql javascript parser that did not make it into the zql.es.js file.

@henridf, in the RenameProc PR brimsec/zq#998 , when you ran "make", did the zql.es.js file not get output? I tried to make it part of the default make tasks but I may have not accounted for all environments.