zq update through "remove file sync from replacer" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1037, authored by @alfred-landrum,
has been merged.

remove file sync from replacer

Replacer is used to update files for zar and zqd. We don't rely on
stability guarantees for the data; we just rely on the atomic properties
of the rename operation. So there's no need to suffer the performance
hit of flushing to disk.

fyi, I verified that this sync (present or not) did not impact the work in https://github.com/brimsec/zq/pull/956 .