zq update through "s3io: strip starting slash from key" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1032, authored by @mattnibs,
has been merged.

s3io: strip starting slash from key

This was causing an issue on the real s3 api for the Prefix property
of the delete batch and list batch api methods. Works for minio's
mock s3 implementation, not for the real s3 api. As it turns out
the starting "/" is not needed for keys, just strip it in
s3io.parsePath.

Closes brimsec/zq#1026 & brimsec/zq#1024