zq update through "force mtime update on metadata file update" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/956, authored by @alfred-landrum,
has been merged.

force mtime update on metadata file update

After some research and experimentation, it seems that the simplest & most reliable way to ensure the mtime is always updated is to directly set it via `os.Chtimes`. I tried an approach of stat'ing the file before the update, then stat'ing it afterwards, but it seems more efficient to just always do so.