zq update through "AST analysis functions to determine source columns" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1033, authored by @henridf,
has been merged.

AST analysis functions to determine source columns

This implements AST analysis to support "column pruning" as described in @alfred-landrum 's Initial Query Planner slides. It is not yet wired up to anything, but the idea is once we have a column-capable scanner,  we can call this during compilation, and pass the resulting list of column names to the scanner.

closes brimsec/zq#1021