zq update through "Perform head call before s3 remove" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1045, authored by @mattnibs,
has been merged.

Perform head call before s3 remove

The s3 api will return 204 on a delete object call whether or not
the object exists. Do a s3 Head call on the object before deleting
to verify that the requested object does in fact exist.

Closes brimsec/zq#1025