zq update through "repo readme update" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1049, authored by @alfred-landrum,
has been merged.

repo readme update

Update the top level readme so that it doesn't just focus on the zq
command.

I also added brief readme's to the directories under `cmd`, to
provide easy to access descriptions for anyone browsing the repo.

Closes brimsec/zq#959