Column Visibility Defaults (#953)
* The default isVisible is based on if all others are visible

* Show All Hide All

Make these their own action types and they now operate on the
entire preferences object, not just the ones that were returned
from the search.

* Column Preferences Touch

When ever we get columns from the search results, always make
a key in the preferences with a sound default at the time.

The logic is, if all are selected, default to visible. If none
exist, default to visible. If some are hidden, default to hidden.

* Use new column pref methods in the app