zq update through "dont buffer terminal output" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1061, authored by @alfred-landrum,
has been merged.

dont buffer terminal output

While working on brimsec/zq#1040 , I was running a locally modified `zar zq` that concurrently executes on archive files, and was surprised that I waited until the entire archive had been searched before seeing results. It turns out that was due to output buffering. This change disables buffering for terminal output, so that results are seen quickly.