zq update through "Disable ZNG compression in zio_test.TestRaw" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1067, authored by @nwt,
has been merged.

Disable ZNG compression in zio_test.TestRaw

zio_test.TestRaw is meant to run without ZNG compression but runs with
it because boomerang ignores its compress parameter.  Fix boomerang to
respect that parameter.

This should have been in brimsec/zq#1050. I overlooked it in my haste to merge.