zq update through "CI test for ZNG output check" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1065, authored by @philrz,
has been merged.

CI test for ZNG output check

See the comments in the script in this PR for the background/motivation.

An example of the test passing with current `zq` and [zq-sample-data](https://github.com/brimsec/zq-sample-data):

https://github.com/brimsec/zq/actions/runs/196891811

I also simulated a failure by creating a branch on which I reverted the commit `6b19b3b` where ZNG compression was enabled, and let this script run against the ZNG data currently in zq-sample-data. Since the current ZNG data in zq-sample-data is compressed but the failure branch was creating uncompressed output, the test failed as shown here:

https://github.com/brimsec/zq/actions/runs/196887459

Before I started writing this test, I'd thought that maybe I'd only let this Action run post-merge since we expect it to fail rarely and it's not a catastrophe if the zq-sample-data is briefly out of date. However, once I saw how it could run as a parallel job, I figured I'd take a shot at letting it run parallel alongside the existing CI. The overall runtime for the CI seems roughly in line with where it was before, and I figure developers might appreciate knowing this info sooner if they introduced some kind of ZNG output bug that wouldn't otherwise have been caught. However, if anyone wants to hold me to my initial sentiment, I won't fight it.