zq update through "create iosrc.Reader to implement ReadAt and Seek" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1068, authored by @mccanne,
has been merged.

create iosrc.Reader to implement ReadAt and Seek

This change extends the reader interface returned by iosrc.Source.NewReader
to allow sub-systems like the micro-indexer to seek and do ReadAt's.
The new interface type, iosrc.Reader, comprises io.Reader, ReaderAt,
Seeker, and Closer.  In particular, this change will allow an iosrc.Reader
to be wrapped in an io.SectionReader.