zq update through "groupby: Add support for partial reducer results" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1077, authored by @henridf,
has been merged.

groupby: Add support for partial reducer results

This PR adds two flags to `ast.GroupbyProc` to indicate that inputs (resp outputs) are partial results from/for decomposable reducers. The flags are not currently exposed in ZQL as they are intended for use by AST transformations (such as those coming in brimsec/zq#1022).

I haven't added tests in this PR and plan to add them with brimsec/zq#1022, unless there are objections.

closes brimsec/zq#1023