zq update through "Add signal-canceled context to commands that call driver.Run" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1099, authored by @nwt,
has been merged.

Add signal-canceled context to commands that call driver.Run

proc.GroupBy and proc.Sort rely on context cancelation for cleanup of
temporary files created by proc.runManager.  Add a signal-canceled
context in pkg/signalctx and use it in commands that call driver.Run
(and thus might create a proc.GroupBy or proc.Sort) to ensure that
cleanup happens.

This in combination with brimsec/zq#1093, this should address brimsec/zq#970.