zq update through "update help text for zar commands" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1094, authored by @alfred-landrum,
has been merged.

update help text for zar commands

We've added support for S3 root's since brimsec/zq#1031 was filed, so I thought I'd add some text both for `zar import` and the higher level `zar help` that mentioned S3. I also tried to make a few other improvements, like consistent language about a "chunk" as the main log file, and the use of a "root" location instead of just directory for the `-R`/`$ZAR_ROOT` configs.

Closes brimsec/zq#1031