zq update through "don't issue empty batches from reduce" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1102, authored by @alfred-landrum,
has been merged.

don't issue empty batches from reduce

I noticed this while debugging a separate issue: reduce's Pull would respond with a batch for each Get() it made, even if it responded with an empty batch. I'm not aware of problems due to empty batches, but it seems more efficienct to loop over Get until end of stream & generate a single non-empty batch, as I see other Procs do.