zq update through "distribute license and acknowledgments file with zq releases" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1107, authored by @alfred-landrum,
has been merged.

distribute license and acknowledgments file with zq releases

Move third party license texts in the zq repo to a single file, 'acknowledgments.txt', and copy it & our license file to the zq repo release assets.