zq update through "K8s" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1101, authored by @marktwallace,
has been merged.

K8s

When reviewing this PR, please work through the steps in k8s/README.md

You can stop at "Adding Observability" -- that part works but is not yet refined.

This PR is additive -- there are no destabilizing changes.