zq update through "impose limit on the number of levels in a microindex" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1119, authored by @mccanne,
has been merged.

impose limit on the number of levels in a microindex

This commit adds a limit on the number of levels that comprise the
index-key tree in a microindex.  Since each node is relatively large
(e.g., 1000s of keys), this limit will only be hit if the frame threshold
is made small enough to limit the number of keys per level to 1.
Even with 2 keys per level, we can represent 2^MaxLevels keys in the
base layer, e.g., a million or so keys, but more typically with
say 512 keys per level, we'd be at 512^20 = 2^200.

Fixes brimsec/zq#995