zq update through "add version check to zdx module" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1120, authored by @mccanne,
has been merged.

add version check to zdx module

This commit changes the microindex version type from a string to
an integer and adds a version check when opening a microindex
that currently fails if the version numbers do not match exactly.

Fixes brimsec/zq#1116