zq update through "Configurable merge" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1121, authored by @henridf,
has been merged.

Configurable merge

Add two fields to ParallelProc to specify ordering behavior when merging branches.

If `ParallelProc.MergeOrderField` is zero-valued, then the existing (unordered) `proc.Merge` proc is used. Otherwise, a `proc.OrderedMerge` proc is used, configured with the ordering field name and direction indicated in the new AST fields.

These new fields are not currently exposed in ZQL syntax but could be at a later point if needed.

closes brimsec/zq#1036