zq update through "Remove reduce proc" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1122, authored by @henridf,
has been merged.

Remove reduce proc

We've discussed in the past how we don't need a separately implemented reduce proc; this PR finally removes it.

(This was prompted by realizing that brimsec/zq#1023 needed to be mirrored for the reduce proc, and figuring that instead this would be a good time to remove it).