zq update through "make MultiFileReader ScannerAble" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1124, authored by @alfred-landrum,
has been merged.

make MultiFileReader ScannerAble

Make a MultiFileReader implement the scanner.ScannerAble interface, so that we can use the scanner speedups on the individual files in a MultiFileReader. While working on the test, I also found a need to replicate some code to copy data from a Pull() like object to a Writer, so I added a `zbuf.Puller` interface and a copy utility.

Needed for https://github.com/brimsec/zq/issues/1075 .