zq update through "fix batch.Unref() bug" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1135, authored by @mccanne,
has been merged.

fix batch.Unref() bug

This commit fixes a bug in archive.FlowgraphIndexer.Write() where
the incoming batch was not unreferenced in an error condition.

This is not that big a deal since not-unref'd stuff would be picked up by GC
and this only happens here on an error, but this is more consistent with the
Unref protocol.