zq update through "zqd: Move pcap storage into its own package" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1126, authored by @mattnibs,
has been merged.

zqd: Move pcap storage into its own package

The idea here is that at a root level a space does not need to know
anything about how pcaps storage is managed. The space provides a
pcapstorage.Store with a location and the Store manages the rest.

Also:
- Add some helpful utilities to iosrc.Source: WriteFile and ReadFile.

closes brimsec/zq#770