zq update through "add support for empty microindexes" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1136, authored by @mccanne,
has been merged.

add support for empty microindexes

This commit adds support for the notion of a valid microindex with
nothing in it.  This is useful for knowing when something has been
indexed but there are no matching elements.  An empty microindex
is a zng file with a microinidex trailer with various null fields.

We also fixed a bug where zar index would leave behind an empty file
(which is not a valid microindex) when it encountered an error.
The fix is to add an Abort() method on the microindex Writer that can
be called in place of Close() to cleanup any partial state.

Fixes brimsec/zq#987