zq update through "Record past allocations in zq memory profile" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1143, authored by @nwt,
has been merged.

Record past allocations in zq memory profile

The -memprofile flag to zq records a sampling of allocations for live
objects upon exit, but that isn't particularly useful.  Record a
sampling of all past allocations instead.