zq update through "add zar map and archive multisource support" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1138, authored by @alfred-landrum,
has been merged.

add zar map and archive multisource support

This splits out the "for each file" like operations of `zar zq` into a new subcommand `zar map`, adds multisource support for archive.Archive, and uses the multisource support in the remaining uses of `zar zq`.

Closes brimsec/zq#1092