zq update through "AST transform for parallelizing a flowgraph" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1145, authored by @henridf,
has been merged.

AST transform for parallelizing a flowgraph

Add `driver.parallelizeFlowgraph()`, a function that tries to transform a sequential flowgraph into a parallel one.

Wiring this into compilation will come in a subsequent PR (I have that working locally).

closes brimsec/zq#1022