zq update through "proc.CompileProc: add support for multiple parents" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1156, authored by @henridf,
has been merged.

proc.CompileProc: add support for multiple parents

Extend proc.CompileProc to support compiling with multiple parents. This will allow passing in multiple parallel scanners into a flowgraph that is suitably shaped (namely, that starts with a parallel proc).

part of brimsec/zq#1155