zq update through "fix bug in put proc and refactor" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1158, authored by @mccanne,
has been merged.

fix bug in put proc and refactor

The put proc had a bug where it attempted to append a clause result
after all of the clauses had already replaced fields in the original
input.  Put.put() was doing many different things with some confusing
variable and type names so a refactor was warranted.

This commit fixes the bug and refactors the logic into bite-sized chunks
with hopefully more helpful names.  The resulting code is smaller.

Fixes brimsec/zq#1152