zq update through "make proc more ergonomic" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1161, authored by @mccanne,
has been merged.

make proc more ergonomic

This commit cleans up the proc package by moving each proc into its
own package.  This change nicely reveals the interdependencies between
the different procs and it removed the proc testing code out of the core
proc package and into proctest.  (Most of the proc tests could be ported
to ztest, which would  be much easier to maintain and encourage more
test writing.)

A proc type that used to be called proc.Foo is now called foo.Proc.

We also changed proc.Base to proc.Parent and simplified the bookkeeping
of the parent linkage with helper methods.