zq update through "Mention Go version in README" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1168, authored by @philrz,
has been merged.

Mention Go version in README

A couple times lately, I've advised community users to compile `master` of `zq` so they could work around a bug before the next GA release. The steps in the README for an install seem simple enough, but the fact that we don't mention the Go version might set them up for trouble if, say, they already have Go installed on their Linux distro and assume that whatever version is there will be adequate. I know we mention the "Go 1.14 or later" requirement in the CONTRIBUTING doc, but since ordinary users just compiling code may not look at that doc, I figure we should be redundant and mention this in the README as well.