zq update through "remove an unncessary waitgroup in log ingester" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1176, authored by @mccanne,
has been merged.

remove an unncessary waitgroup in log ingester

This commit removes an unncessary waitgroup from the log ingester
since synchronous was already being provided by the closure of
the warnings channel.  We also added comments for clarification.
This follows more closely the pattern of context.Context.