zq update through "zqd: prevent being orphaned by Brim (unix only)" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1184, authored by @mattnibs,
has been merged.

zqd: prevent being orphaned by Brim (unix only)

Add hidden -brimfd flag to zqd listen. Brim will use this option
to prevent zqd from being orphaned in the event that Brim is terminated
with a SIGKILL signal.

If set zqd with listen to the file descriptor and gracefully should
it be closed.

PART OF brimsec/brim#1018