zq update through "use fewer api calls in s3 reader" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1191, authored by @alfred-landrum,
has been merged.

use fewer api calls in s3 reader

Use far fewer api calls to service Reads for an s3io.Reader by using a single GetObject request when the file is first read or after a Seek changes the offset.