zq update through "zqd: Gracefully handle moved/delete pcap in spaces" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1198, authored by @mattnibs,
has been merged.

zqd: Gracefully handle moved/delete pcap in spaces

Currently a space that references a moved or delete pcap causes
the list spaces endpoint to fail with 404.

If a pcapspace references a no longer existing pcap make it so:
- list spaces still works
- info request on the space returns pcap_support false
- searches on the space still works

Closes brimsec/zq#1195