zq update through "K8s obs" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1173, authored by @marktwallace,
has been merged.

K8s obs

This updates the procedures and config files for EKS deployment of zqd.
It also adds tools for observability, including an example Grafana dashboard for zqd.
The PR does not yet expose EKS endpoints, so kubectl port-forward must be used for testing.