zq update through "Parallelize further when tail is order-insensitive" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1194, authored by @henridf,
has been merged.

Parallelize further when tail is order-insensitive

If the "tail" of a parallelized flowgraph doesn't care about input order (for example, if the flowgraph ends with ` | count()`), then we can still parallelize procs that modify the ordering field.

closes brimsec/zq#1034