zq update through "Parsimonious use of zng.Record.TypeCheck()" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1192, authored by @henridf,
has been merged.

Parsimonious use of zng.Record.TypeCheck()

We currently run TypeCheck() on every record read by a zio reader,
which for some workloads adds significant overhead (see brimsec/zq#1181). Yet
for data that we've imported and validated along the
way (specifically, zar and zqd), this check is unneeded.

This commit addresses the above by making the TypeCheck() configurable
via a flag in detector.OpenConfig.

zar import and zqd log ingest enable it. zq enables it and also
exposes the knob through a new `-check` CLI option. It is disabled in
other uses, such as `zar zq` or zqd queries.

closes brimsec/zq#1181