zq update through "Handle nested records with nonadjacent fields in zeekio.Parser" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1225, authored by @nwt,
has been merged.

Handle nested records with nonadjacent fields in zeekio.Parser

zeekio.Parser assumes the fields of a nested record are adjacent.  Relax
that requirement.

I find this hard to love but couldn't come up with anything nicer.

Should close brimsec/zq#1217.