zq update through "cut -c: use descriptor to derive fields" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1227, authored by @henridf,
has been merged.

cut -c: use descriptor to derive fields

Previously, cut -c used a zng.fieldIter on the first record of a type
to compute which fields to extract. This leads to problems with unset
values: if an inner record is unset, fieldIter doesn't visit leaves,
and the cutter gets confused, as observed in https://github.com/brimsec/zq/issues/1226.

Instead, use the record's descriptor to figure out which fields to cut.

closes brimsec/zq#1226