zq update through "Add -sortmem flag to zar import, and a yaml test" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1203, authored by @marktwallace,
has been merged.

Add -sortmem flag to zar import, and a yaml test

This adds the same -sortmem flag that is used in zq to the zar export command. It writes an error to stderr if the sort space is smaller the 1 MB, which I determined to be a reasonable lower limit. There is a  ztest yaml test with a positive and negative test case for -sortmem.