zq update through "replace *zio.Writer with zbuf.WriteCloser" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1236, authored by @mccanne,
has been merged.

replace *zio.Writer with zbuf.WriteCloser

This commit cleans up the zio design pattern to allow generic
writers to be closed instead of doing a flush followed by reaching
around the writer to close the underlyinig object.

This will make integrating the zst writer easier because it needs
to be closed to finish up writing metadata and a trailer.  I was
fighting with the existing pattern to wire it all up and this change
will make it all straightforward.

The previous flush-based approach had merits but this seems
easier to understand and work with, albeit we had to add Close
stubs to all of the zio writers.

We also mooved zbuf/zng.go to zbuf/zbuf.go as the previous
name was historical and no longer accurate, and deleted
zbuf.WriteFlusher as it is no longer used.

Finally, this refactoring exposed a problem in the fs.ReplaceFile
call in storage filestore, which was doing a storage copy without
any write buffering, so this is fixed.

Fixed brimsec/zq#1235