zq update through "move zng flattener from zeekio to zng/flattener" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1252, authored by @mccanne,
has been merged.

move zng flattener from zeekio to zng/flattener

This commit moves the flattening logic out of the zeekio package
and into its own package under zng since this operation is a
generic zng transformation and is not specific to zio.

This will be helpful also to the csv writer when we add it.