zq update through "simplify zng visitor" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1258, authored by @mccanne,
has been merged.

simplify zng visitor

I was working on updating zng sets and didn't understand why I had
to change so many places in zng/walk.go so I simplified it.

This commit simplifies the code in zng/walk.go to use a common
function for descending into container types.  The error messages
are a little different and a union record value gets visited now
instead of omitted, but I think that was a bug/oversight.

We also changed zng.RecordVisitor to zng.Visitor since you can now
called zng.Walk on any type.