zq update through "fix zqd version breakage" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1256, authored by @mccanne,
has been merged.

fix zqd version breakage

This commit fixes a bug introduced by the recent refactoring of
cli flags and zio options.  While we're at it, we added a zapi
version command and added a test to make sure it's working.

We also simplified the format of the version response message.

The brim app doesn't seem to use the /version endpoint so the
change shouldn't break anything.

Fixes brimsec/zq#1255