zq update through "update zapi to package zq/cli and other minor changes" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1278, authored by @mccanne,
has been merged.

update zapi to package zq/cli and other minor changes

This commit updates "zapi get" to use the output flags options
in package zq/cli.  It also changes the "protocol" option to
"encoding" representing the server format encoding of the http
response.  This sets up "zapi get" to handle the vanilla ndjson
and csv support that is coming soon to the zqd search end point.

We also updated the online help a bit.