zq update through "Add test for nanosecond pcap" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1241, authored by @philrz,
has been merged.

Add test for nanosecond pcap

Since https://github.com/brimsec/brim/pull/1069 seeks to make nanosecond pcap one of the supported formats we let in via the app, I figured we should have a test in the backend that does a basic smoke test with the format.