zq update through "add csv option to search endpoint" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1280, authored by @mccanne,
has been merged.

add csv option to search endpoint

This commit adds a csv option to the search endpoint enabled by
setting the "format" query parameter on the search to "csv".

We also fixed a small problem in "zapi get" so that "-e csv"
will work properly.