zq update through "add json option to search endpoint" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1285, authored by @mccanne,
has been merged.

add json option to search endpoint

This commit renames the old json search endpoint "zjson" and adds
a new json format option that returns a valid, one-shot json
representation of query results.  This is enabled by
setting the "format" query parameter on the search to "json".

Fixes brimsec/zq#1277