zq update through "Mention icmp as a proto" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1281, authored by @philrz,
has been merged.

Mention icmp as a proto

While working on brimsec/zq#1240 I noticed that `-p icmp` is a legal option but it's not currently mentioned in the `pcap help slice` guidance. Since this brings us from two options to three, I opted to use the `[one,two,three]` syntax similar to how the input/output formats are presented in the `zq` / `zar` help.