zq update through "import joe and make it better" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1284, authored by @mccanne,
has been merged.

import joe and make it better

This commit rewrites joe to be more go-like and better.

Many have complained about joe, me included, but I still think
it's easier to maintain code written with this thin API compared
to seeing []interface{} and map[string]interface{} all over
the place.

p.s. I was using the old joe for the zjson changes and got frustrated
so I did this and it felt better.