zq update through "Point to permanent link for joining public Slack" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1304, authored by @philrz,
has been merged.

Point to permanent link for joining public Slack

We recently added a redirect layer on the https://www.brimsecurity.com/ web site such that the `#` icon for joining Slack is attached to a permanent URL https://www.brimsecurity.com/join-slack/ which, in turn, redirects to the current "invite yourself" URL, which might change after 2000 users invite themselves or Slack expires it due to some glitch (it's happened before). Here I change all the Slack URL pointers such that if the "invite yourself" URL changes again in the future, we only have to update it in that one place on the web site.