zq update through "simplify resolver.Slice" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1308, authored by @mccanne,
has been merged.

simplify resolver.Slice

This commit simplifies the resolver.Slice type so it's just a
slice of *zng.Records instead of an embedded struct.  This way
it can be transparently used as this type outside of the resolver
module.  The upcoming dataframe proc will use this.