zq update through "create zst format" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1268, authored by @mccanne,
has been merged.

create zst format

This commit introduces the zst columnar format to zq.  Package zst
provides the core new functionality.  Package zstio is a zio wrapper
for zst to wire it into the zq pipeline.  A new CLI command "zst"
provides a tool for test, debug, and demo.

There is much more work to be done for columnar format but this
should serve as a good foundation to start from.

Fixes brimsec/zq#1219