zq update through "align zapi with rest of tools that print help with no args" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1328, authored by @mccanne,
has been merged.

align zapi with rest of tools that print help with no args

This commit changes "zapi" to print help when run with no args.
The repl is now run with a new subcommand "zapi repl".

Fixes brimsec/zq#1306