zq update through "pcapng: fix bad parse interface bug" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1341, authored by @mattnibs,
has been merged.

pcapng: fix bad parse interface bug

The parseInterfaceDescriptor was not advancing the read buffer
before reading through the list of options for the block. This
was causing a red-herring bad-option-length error; the reader was
not actually reading the option list block.

closes brimsec/zq#1057