zq update through "Use Brim Suricata package (macOS, Ubuntu)" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1331, authored by @henridf,
has been merged.

Use Brim Suricata package (macOS, Ubuntu)

This change adjusts our Makefile and a few related files to
automatically download and use the Brim suricata package when running
on macOS or Ubuntu. The process is overall similar to the one for zeek.

This also adjusts one system test to ensure that suricata ingest
completes properly.