zq update through "Fix wildcard field name example by introducing **" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1337, authored by @philrz,
has been merged.

Fix wildcard field name example by introducing **

I was re-reading the docs with an eye to making improvements, but I stumbled onto this place where we regressed. The tl;dr is that this doc example stopped working with the changes in brimsec/zq#59, but no one caught it because it pre-dated Michael Brown's automation that ensured docs examples stayed in sync with `zq` outputs against out sample data. Then in brimsec/zq#330 where Michael introduced that automation, he pasted in updated `zq` output that made the test pass, but the output no longer made sense with the accompanying text. It's been that way ever since.

Figuring out how to fix it was a trip down memory lane. I'd forgotten how around the time of brimsec/zq#59, Andrew changed the behavior of `*=` such that it only matched against _top-level_ fields, and then introduced `**=` that also targeted fields inside of nested records. I had a long 1-on-1 Slack thread going with Andrew at the time about how I didn't love this change, but I guess conditions were such that I never argued the prior behavior back into existence nor did `**=` end up being documented. So I do my best to reconcile all that here.

I admit that it's not the easiest thing to read, but I feel like that's a reflection of the underlying complexity of the operators. If looked at as a likely `jq` substitute, I respect that having operators that target "levels" of nesting is sometimes desirable. OTOH for the security-centric use cases that the Brim app has in mind (and these ZQL docs were written for an audience of those security users, which is why Zeek data is used in the examples) I have difficulty imagining real world examples where it matters. All that said, the text & example before this PR were confusing & wrong, so it seems I'd have to make a choice between fixing it or not mentioning `*=` / `**=` at all, and I've chosen to attempt the former. If there's a silver lining, no _users_ have complained about this doc, which says to me that this may be power user functionality that users aren't reaching for yet, so we probably don't need to overthink it regardless.

I know we have a high-level to-do to polish more corner cases in ZQL (brimsec/zq#861), so perhaps this is one to consider during that exercise.