zq update through "pcap: Add pcap info command" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1354, authored by @mattnibs,
has been merged.

pcap: Add pcap info command

zq#1341 fixed a bug with how pcapio was parsing interface descriptors
for pcapng files. This follow up as an info command along with tests
to ensure the pcap command is correctly gathering pcap info.

This is a simple first pass, not an attempt to exhaustively parse and
display all information available in the pcap|ng file formats.