zq update through "allow ztest yaml tests to live anywhere" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1160, authored by @mccanne,
has been merged.

allow ztest yaml tests to live anywhere

This PR proposes an extension to the system tests to allow yaml ztests to be put anywhere in the source tree.  This would allow us to use ztest to more easily write unit-like tests throughout the code.

For example, we can now move a test like this, which is really a unit test for the put proc,
```
tests/suite/put/put-1.yaml
```
to
```
proc/put/ztests/put-1.yaml
```
The logic currently looks for any yaml files that are a descendent of a directory called `ztests` or any yaml files under the top-level `tests` directory.  I was thinking we could move `tests/suite` and `tests/formats` to a new top-level directory `ztests` and get rid of the `tests` special case maybe do this in a subsequent PR?