zq update through "fix record volatility bug introduced in commit 0382a0f" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1362, authored by @mccanne,
has been merged.

fix record volatility bug introduced in commit 0382a0f

This commit fixes a bug that was introduced in commit 0382a0f,
where a rec.Keep() call was dropped in the spill/read path compared
to the old code.  This fix puts the call back in.

Fixes brimsec/zq#1359