zq update through "support overlapping data files in an archive" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1348, authored by @alfred-landrum,
has been merged.

support overlapping data files in an archive

When the data files for a query have overlapping time spans, scan & read the overlapping files, merging their records in the archive's order.

On top of brimsec/zq#1330
Closes brimsec/zq#1263