zq update through "Allow for singular "week" like we do for other time units" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1374, authored by @philrz,
has been merged.

Allow for singular "week" like we do for other time units

While documenting aggregate functions, I noticed that the singular `week` without a leading number was not available, though it is for all the other supported time units. I couldn't see a reason for leaving this out, so I've followed the existing pattern to make this consistent.