zq update through "backward compat for reading the alpha version of zng files" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1386, authored by @mccanne,
has been merged.

backward compat for reading the alpha version of zng files

THIS IS A SHORT COMMIT.  DON'T LET THE LINE COUNT FOOL YOU.
THERE ARE TWO COMMITS.  YOU ONLY NEED TO REVIEW THE SECOND ONE.

The commit is only big because we're copying a snapshot of the current
zng package and couple other things to retain backward compat.  We will remove
this later when zng-alpha is fully deprecated.  Please review the second commit
in the PR.   The first commit is the snapshot of the zng package and the code
is completely unmodified except for some of the import paths to have them
refer to zq/alpha/{zng,zio} instead of zq/{zng,zio}.

===

This commit adds backward compatibility for reading alpha-version
zng files.  Right now, the tip of the tree is still the alpha version
but we will soon introduce changes so we are putting this backward
compat in now, in front of coming changes.

This commit introduces a copy of package zng and other bits we need
under the zq/alpha tree.  This can be removed down the road when we
fully deprecate the alpha zng format.

Since the tzng format hasn't changed except for "byte" to "uint8",
which isn't currently easy to get into a zng file without hand editing,