zq update through "fix bug in spill file name allocation" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1391, authored by @mccanne,
has been merged.

fix bug in spill file name allocation

The recent refactor of the spill code into its own module introduced
a bug where the file names could collide because they were named
after their position in the spill table, which grows and shrinks.
This caused distinct spills to have the same file name.  The fix is
to track the total number of spills and name the file after this.