zq update through "add spilling to fuse proc" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1355, authored by @mccanne,
has been merged.

add spilling to fuse proc

This commit adds spill logic to the fuse proc triggered when its memory
footprint exceeds a threshold.  We changed the fuse proc implementation
to use a goroutine so it can cleanly teardown the spill file when
the proc context is canceled in addition to providing parallelism
between scanning and spilling.

The -fusemem proc flag option is analogous to -sortmem and controls
how much memory fuse will use before spilling.

We also added a Read method in zbuf.Array to implement zbuf.Reader,
which is handy for turning a slice of zng records into a reader.

Closes brimsec/zq#1270