zq update through "JSON types: remove "_path must be first column" check" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1370, authored by @henridf,
has been merged.

JSON types: remove "_path must be first column" check

This commit removes the validation check that ensured that _path was
present as the first field of descriptors. This check is a reflection of our
zeek-onlyness back when this feature was first introduced. Now with the
impending arrival of Suricata and other data types, this hardcoded
check no longer belongs.

If we find ourselves needing such a validation, we can re-introduce
something more general once we have a few more ingest types under our
belt.