zq update through "Aggregate function & grouping docs" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1385, authored by @philrz,
has been merged.

Aggregate function & grouping docs

Here's the long overdue ZQL docs on aggregate functions & related grouping concepts. The docs and examples are based on my own black box testing and attempts to interpret the grammar. There might be subtleties that I've not captured, so please speak up if you think something is missing.

I also recognize that some of this might be subject to change as @mccanne continues his overhaul of expressions in particular and ZQL in general. However, as I've had community members recently cite the existence of these parts of the language and ask for guidance on use, I figure it's better to capture it & iterate rather than continuing to kick the can down the road.