zq update through "Convert links from godoc.org to pkg.go.dev" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1401, authored by @philrz,
has been merged.

Convert links from godoc.org to pkg.go.dev

I recently noticed the text at the top of godoc.org:

![image](https://user-images.githubusercontent.com/5934157/94733865-431f6d00-031d-11eb-975e-24caee9f2a04.png)

I don't have a strong opinion on old vs. new, but since the message seems to imply an active transition, this PR includes the changes if we wanted to embrace it. The one link to godoc.org that it seems can't be changed is this one in the top-level README:

```
[doc-img]: https://godoc.org/github.com/brimsec/zq?status.svg
```

That link seems to go to the same destination page as https://pkg.go.dev/github.com/brimsec/zq, so I've left it as-is.